requirements.txt
flask
flask-cors
torch
opencv-python
Pillow
